Move uploadFileOnLink to link module.
